Exit if release branch not found

+++ src/circleci-base/scripts/release-abort.sh
@@ -5,6 +5,11 @@ git flow release
 
 release=${1:-$(git flow release | head | tr -s ' ' | cut -d' ' -f2)}
 
+if [[ -z "$release" ]]
+then
+  exit 1
+fi
+
 git reset --hard
 
 git checkout ${CIRCLE_SHA1:-develop}
